Fix solenoid API

+++ src/main/include/SolenoidManager.hpp
@@ -14,19 +14,19 @@ class SolenoidManager : public Singleton<SolenoidManager>
 {
     friend Singleton;
 public:
-    void registerSolenoid(ck::SolenoidControl::Solenoid::ModuleType moduleType, uint16_t id, ck::SolenoidControl::Solenoid::SolenoidType solenoidType);
-    void deleteSolenoid(uint16_t id);
-    void onSolenoid(uint16_t id, std::function<void(uint16_t, CKSolenoid*, ck::SolenoidControl::Solenoid::SolenoidType)> func);
-    void onSolenoid(const google::protobuf::Message& msg, std::function<void(uint16_t, CKSolenoid*, ck::SolenoidControl::Solenoid::SolenoidType, const ck::SolenoidControl::Solenoid&)> func);
-    void forEach(std::function<void(uint16_t, CKSolenoid*, ck::SolenoidControl::Solenoid::SolenoidType)> func);
+    void registerSolenoid(ck::SolenoidControl::Solenoid::ModuleType moduleType, uint32_t id, ck::SolenoidControl::Solenoid::SolenoidType solenoidType);
+    void deleteSolenoid(uint32_t id);
+    void onSolenoid(uint32_t id, std::function<void(uint32_t, CKSolenoid*, ck::SolenoidControl::Solenoid::SolenoidType)> func);
+    void onSolenoid(const google::protobuf::Message& msg, std::function<void(uint32_t, CKSolenoid*, ck::SolenoidControl::Solenoid::SolenoidType, const ck::SolenoidControl::Solenoid&)> func);
+    void forEach(std::function<void(uint32_t, CKSolenoid*, ck::SolenoidControl::Solenoid::SolenoidType)> func);
     void processHeartbeat();
-    CKSolenoid* getSolenoid(uint16_t id);
+    CKSolenoid* getSolenoid(uint32_t id);
 private:
     SolenoidManager();
     ~SolenoidManager();
-    std::map<uint16_t, CKSolenoid*> mRegisteredSolenoidList;
-    std::map<uint16_t, ck::SolenoidControl::Solenoid::SolenoidType> mRegisteredSolenoidTypeList;
-    std::map<uint16_t, int> mRegisteredSolenoidHeartbeatList;
+    std::map<uint32_t, CKSolenoid*> mRegisteredSolenoidList;
+    std::map<uint32_t, ck::SolenoidControl::Solenoid::SolenoidType> mRegisteredSolenoidTypeList;
+    std::map<uint32_t, int> mRegisteredSolenoidHeartbeatList;
 
     std::recursive_mutex solenoidLock;
